requirements.txt
# Web Framework
gradio==4.19.2

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.2.0

# AI/ML
google-generativeai==0.3.2
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.1.2

# NLP & Embeddings
scikit-learn==1.4.0
numpy==1.26.3

# Utilities
python-dotenv==1.0.1
